[수정] 자동 생성 계정 추가 #160
수정사항: 자동 생성 계정 추가

영향 범위: 없음
